refactoring views.py
view classes and functions for models Project, Task, Team, Worker separated into separate files
