+++ task_manager/views/team_view.py
@@ -1,0 +1,34 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.urls import reverse_lazy
+from django.views import generic
+
+from task_manager.forms import TeamForm
+from task_manager.models import Team
+
+
+class TeamDetailView(LoginRequiredMixin, generic.DetailView):
+    model = Team
+
+
+class TeamListView(LoginRequiredMixin, generic.ListView):
+    model = Team
+    context_object_name = "team_list"
+    template_name = "task_manager/team_list.html"
+    paginate_by = 20
+
+
+class TeamCreateView(LoginRequiredMixin, generic.CreateView):
+    model = Team
+    form_class = TeamForm
+    success_url = reverse_lazy("task_manager:team-list")
+
+
+class TeamUpdateView(LoginRequiredMixin, generic.UpdateView):
+    model = Team
+    form_class = TeamForm
+    success_url = reverse_lazy("task_manager:team-list")
+
+
+class TeamDeleteView(LoginRequiredMixin, generic.DeleteView):
+    model = Team
+    success_url = reverse_lazy("task_manager:team-list")
